chore: rename release and build paths
