+++ .erb/scripts/check-native-dep.js
@@ -30,14 +30,14 @@ if (dependencies) {
  )}
 ${chalk.bold(filteredRootDependencies.join(', '))} ${
         plural ? 'are native dependencies' : 'is a native dependency'
-      } and should be installed inside of the "./build/app" folder.
+      } and should be installed inside of the "./release/app" folder.
  First, uninstall the packages from "./package.json":
 ${chalk.whiteBright.bgGreen.bold('npm uninstall your-package')}
  ${chalk.bold(
    'Then, instead of installing the package to the root "./package.json":'
  )}
 ${chalk.whiteBright.bgRed.bold('npm install your-package')}
- ${chalk.bold('Install the package to "./build/app/package.json"')}
+ ${chalk.bold('Install the package to "./release/app/package.json"')}
 ${chalk.whiteBright.bgGreen.bold('cd ./src && npm install your-package')}
  Read more about native dependencies at:
 ${chalk.bold(
