+++ .vscode/settings.json
@@ -13,7 +13,7 @@
     ".git": true,
     ".eslintcache": true,
     ".erb/dll": true,
-    "build/{release,app/dist}": true,
+    "release/{build,app/dist}": true,
     "node_modules": true,
     "npm-debug.log.*": true,
     "test/**/__snapshots__": true,
